fixed some options for enums

+++ src/models/s2s.h
@@ -36,7 +36,7 @@ public:
 
     auto rnnFw = rnn::rnn(graph)                                   //
         ("type", opt<std::string>("enc-cell"))                     //
-        ("direction", forward)                                     //
+        ("direction", (int)forward)                                //
         ("dimInput", embeddings->shape()[-1])                      //
         ("dimState", opt<int>("dim-rnn"))                          //
         ("dropout", dropoutRnn)                                    //
@@ -62,7 +62,7 @@ public:
 
     auto rnnBw = rnn::rnn(graph)                                   //
         ("type", opt<std::string>("enc-cell"))                     //
-        ("direction", backward)                                    //
+        ("direction", (int)backward)                               //
         ("dimInput", embeddings->shape()[-1])                      //
         ("dimState", opt<int>("dim-rnn"))                          //
         ("dropout", dropoutRnn)                                    //
@@ -264,7 +264,7 @@ public:
           mlp::dense(graph)                                          //
           ("prefix", prefix_ + "_ff_state")                          //
           ("dim", opt<int>("dim-rnn"))                               //
-          ("activation", mlp::act::tanh)                             //
+          ("activation", (int)mlp::act::tanh)                        //
           ("layer-normalization", opt<bool>("layer-normalization"))  //
           ("nematus-normalization",
            options_->has("original-type")
@@ -327,7 +327,7 @@ public:
     auto layer1 = mlp::dense(graph)                                //
         ("prefix", prefix_ + "_ff_logit_l1")                       //
         ("dim", opt<int>("dim-emb"))                               //
-        ("activation", mlp::act::tanh)                             //
+        ("activation", (int)mlp::act::tanh)                        //
         ("layer-normalization", opt<bool>("layer-normalization"))  //
         ("nematus-normalization",
          options_->has("original-type")
